Interim update of the Development Branch with last improvements

+++ src/main/java/de/fraunhofer/isst/configmanager/controller/ResourceUIController.java
@@ -1,18 +1,14 @@
 package de.fraunhofer.isst.configmanager.controller;
 
-import de.fraunhofer.iais.eis.BaseConnector;
-import de.fraunhofer.iais.eis.Resource;
-import de.fraunhofer.iais.eis.ResourceImpl;
 import de.fraunhofer.iais.eis.ids.jsonld.Serializer;
 import de.fraunhofer.isst.configmanager.communication.clients.DefaultConnectorClient;
-import de.fraunhofer.isst.configmanager.configmanagement.service.ConfigModelService;
 import de.fraunhofer.isst.configmanager.configmanagement.service.ResourceService;
+import de.fraunhofer.isst.configmanager.util.ValidateApiInput;
 import io.swagger.v3.oas.annotations.tags.Tag;
+import lombok.AccessLevel;
+import lombok.experimental.FieldDefaults;
 import lombok.extern.slf4j.Slf4j;
-import net.minidev.json.JSONArray;
 import net.minidev.json.JSONObject;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -22,7 +18,6 @@ import org.springframework.web.bind.annotation.RestController;
 import java.io.IOException;
 import java.net.URI;
 import java.util.ArrayList;
-import java.util.List;
 
 /**
  * The controller class implements the ResourceUIApi and offers the possibilities to manage
@@ -31,21 +26,18 @@ import java.util.List;
 @RestController
 @RequestMapping("/api/ui")
 @Slf4j
-@Tag(name = "Resource Management", description = "Endpoints for managing the resource in the configuration manager")
+@Tag(name = "Resource Management", description = "Endpoints for managing the resource in the " +
+        "configuration manager")
+@FieldDefaults(makeFinal = true, level = AccessLevel.PRIVATE)
 public class ResourceUIController implements ResourceUIApi {
-
-    private final static Logger logger = LoggerFactory.getLogger(ResourceUIController.class);
-
-    private final ResourceService resourceService;
-    private final ConfigModelService configModelService;
-    private final DefaultConnectorClient client;
-    private final Serializer serializer;
+    transient ResourceService resourceService;
+    transient DefaultConnectorClient client;
+    transient Serializer serializer;
 
     @Autowired
-    public ResourceUIController(ResourceService resourceService, ConfigModelService configModelService,
-                                DefaultConnectorClient client, Serializer serializer) {
+    public ResourceUIController(final ResourceService resourceService,
+                                final DefaultConnectorClient client, final Serializer serializer) {
         this.resourceService = resourceService;
-        this.configModelService = configModelService;
         this.client = client;
         this.serializer = serializer;
     }
@@ -57,16 +49,23 @@ public class ResourceUIController implements ResourceUIApi {
      * @return a suitable http response depending on success
      */
     @Override
-    public ResponseEntity<String> getResource(URI resourceId) {
+    public ResponseEntity<String> getResource(final URI resourceId) {
         log.info(">> GET /resource resourceId: " + resourceId);
 
-        Resource resource = resourceService.getResource(resourceId);
+        if (ValidateApiInput.notValid(resourceId.toString())) {
+            return ResponseEntity.badRequest().body("All validated parameter have undefined as " +
+                    "value!");
+        }
+
+        final var resource = resourceService.getResource(resourceId);
 
         if (resource != null) {
             try {
                 return ResponseEntity.ok(serializer.serialize(resource));
             } catch (IOException e) {
-                return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Could not serialize resource!");
+                log.error(e.getMessage(), e);
+                return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Could not " +
+                        "serialize resource!");
             }
         } else {
             return ResponseEntity.badRequest().body("Could not determine the resource");
@@ -81,37 +80,13 @@ public class ResourceUIController implements ResourceUIApi {
     @Override
     public ResponseEntity<String> getResources() {
         log.info(">> GET /resources");
-
-        ArrayList<Resource> resources = resourceService.getResources();
-        if (resources != null) {
-            try {
-                return ResponseEntity.ok(serializer.serialize(resources));
-            } catch (IOException e) {
-                logger.error(e.getMessage());
-                return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Could not serialize resources!");
-            }
-        } else {
-            return ResponseEntity.badRequest().body("Could not determine the resources");
-        }
+        return ResponseEntity.ok(resourceService.getOfferedResourcesAsJsonString());
     }
 
     @Override
     public ResponseEntity<String> getRequestedResources() {
         log.info(">> GET /resources/requested");
-
-        try {
-            BaseConnector baseConnector = client.getSelfDeclaration();
-            if (baseConnector != null) {
-                List<Resource> resourceList = resourceService.getRequestedResources();
-                return ResponseEntity.ok(serializer.serialize(resourceList));
-            } else {
-                return ResponseEntity.badRequest().body("Could not get the self declaration from the connector");
-            }
-        } catch (IOException e) {
-            logger.error(e.getMessage(), e);
-            return ResponseEntity.badRequest().body("Problems occurred while determining the requested resources " +
-                    "from the connector");
-        }
+        return ResponseEntity.ok(resourceService.getRequestedResourcesAsJsonString());
     }
 
     /**
@@ -121,12 +96,17 @@ public class ResourceUIController implements ResourceUIApi {
      * @return a suitable http response depending on success
      */
     @Override
-    public ResponseEntity<String> getResourceInJson(URI resourceId) {
+    public ResponseEntity<String> getResourceInJson(final URI resourceId) {
         log.info(">> GET /resource/json resourceId: " + resourceId);
 
-        Resource resource = resourceService.getResource(resourceId);
+        if (ValidateApiInput.notValid(resourceId.toString())) {
+            return ResponseEntity.badRequest().body("All validated parameter have undefined as " +
+                    "value!");
+        }
 
-        JSONObject resourceJson = new JSONObject();
+        final var resource = resourceService.getResource(resourceId);
+
+        final var resourceJson = new JSONObject();
         resourceJson.put("title", resource.getTitle().get(0).getValue());
         resourceJson.put("description", resource.getDescription().get(0).getValue());
         resourceJson.put("keyword", resource.getKeyword());
@@ -138,30 +118,38 @@ public class ResourceUIController implements ResourceUIApi {
     }
 
     /**
-     * This method deletes the resource from the connector and the app route with the given parameter.
+     * This method deletes the resource from the connector and the app route with the given
+     * parameter.
      * If both are deleted the dataspace connector is informed about the change.
      *
      * @param resourceId id of the resource
      * @return http response from the target connector
      */
     @Override
-    public ResponseEntity<String> deleteResource(URI resourceId) {
+    public ResponseEntity<String> deleteResource(final URI resourceId) {
         log.info(">> DELETE /resource resourceId: " + resourceId);
+
+        if (ValidateApiInput.notValid(resourceId.toString())) {
+            return ResponseEntity.badRequest().body("All validated parameter have undefined as " +
+                    "value!");
+        }
+
         try {
-            var response = client.deleteResource(resourceId);
+            final var response = client.deleteResource(resourceId);
             resourceService.deleteResourceFromAppRoute(resourceId);
-            var jsonObject = new JSONObject();
+            final var jsonObject = new JSONObject();
             jsonObject.put("connectorResponse", response);
             jsonObject.put("resourceID", resourceId.toString());
             return ResponseEntity.ok(jsonObject.toJSONString());
         } catch (IOException e) {
-            e.printStackTrace();
+            log.error(e.getMessage(), e);
             return ResponseEntity.badRequest().body("Could not send delete request to connector");
         }
     }
 
     /**
-     * This method creates a resource with the given parameters. The special feature here is that the created resource
+     * This method creates a resource with the given parameters. The special feature here is that
+     * the created resource
      * is included once in the app route and once in the resource catalog of the connector.
      *
      * @param title           title of the resource
@@ -174,31 +162,44 @@ public class ResourceUIController implements ResourceUIApi {
      * @return response from the target connector
      */
     @Override
-    public ResponseEntity<String> createResource(String title, String description, String language,
-                                                 ArrayList<String> keywords, String version, String standardlicense,
-                                                 String publisher) {
-        log.info(">> POST /resource title: " + title + " description: " + description + " language: " + language + " keywords: " + keywords + " version: " + version + " standardlicense: " + standardlicense
-        + " publisher: " + publisher);
+    public ResponseEntity<String> createResource(final String title, final String description,
+                                                 final String language,
+                                                 final ArrayList<String> keywords,
+                                                 final String version, final String standardlicense,
+                                                 final String publisher) {
+        log.info(">> POST /resource title: " + title + " description: " + description + " " +
+                "language: " + language + " keywords: " + keywords + " version: " + version + " " +
+                "standardlicense: " + standardlicense
+                + " publisher: " + publisher);
 
-        ResourceImpl resource = resourceService.createResource(title, description, language, keywords,
+        if (ValidateApiInput.notValid(title, description, language, version, standardlicense,
+                publisher)) {
+            return ResponseEntity.badRequest().body("All validated parameter have undefined as " +
+                    "value!");
+        }
+
+
+        final var resource = resourceService.createResource(title, description, language,
+                keywords,
                 version, standardlicense, publisher);
 
         // Save and send request to dataspace connector
-        var jsonObject = new JSONObject();
+        final var jsonObject = new JSONObject();
         try {
             jsonObject.put("resourceID", resource.getId().toString());
-            var response = client.registerResource(resource);
+            final var response = client.registerResource(resource);
             jsonObject.put("connectorResponse", response);
             return ResponseEntity.ok(jsonObject.toJSONString());
         } catch (IOException e) {
             jsonObject.put("message", "Could not register resource at connector");
-            logger.error(e.getMessage());
+            log.error(e.getMessage(), e);
             return ResponseEntity.badRequest().body(jsonObject.toJSONString());
         }
     }
 
     /**
-     * This method updates a resource with the given parameters. The special feature here is that the resource
+     * This method updates a resource with the given parameters. The special feature here is that
+     * the resource
      * is updated once in the app route and once in the resource catalog of the connector.
      *
      * @param resourceId      id of the resource
@@ -212,29 +213,39 @@ public class ResourceUIController implements ResourceUIApi {
      * @return response from the target connector
      */
     @Override
-    public ResponseEntity<String> updateResource(URI resourceId, String title, String description, String language,
-                                                 ArrayList<String> keywords, String version, String standardlicense,
-                                                 String publisher) {
-        log.info(">> PUT /resource title: " + title + " description: " + description + " language: " + language + " keywords: " + keywords + " version: " + version + " standardlicense: " + standardlicense
+    public ResponseEntity<String> updateResource(final URI resourceId, final String title,
+                                                 final String description, final String language,
+                                                 final ArrayList<String> keywords,
+                                                 final String version, final String standardlicense,
+                                                 final String publisher) {
+        log.info(">> PUT /resource title: " + title + " description: " + description + " language" +
+                ": " + language + " keywords: " + keywords + " version: " + version + " " +
+                "standardlicense: " + standardlicense
                 + " publisher: " + publisher);
 
-
+        if (ValidateApiInput.notValid(resourceId.toString(), title, description, language, version, standardlicense, publisher)) {
+            return ResponseEntity.badRequest().body("All validated parameter have undefined as " +
+                    "value!");
+        }
 
         // Save the updated resource and update the resource in the dataspace connector
         try {
-            ResourceImpl updatedResource = resourceService.updateResource(resourceId, title, description, language, keywords,
+            final var updatedResource = resourceService.updateResource(resourceId, title,
+                    description, language, keywords,
                     version, standardlicense, publisher);
-            if(updatedResource != null){
-                var response = client.updateResource(resourceId, updatedResource);
+            if (updatedResource != null) {
+                final var response = client.updateResource(resourceId, updatedResource);
                 resourceService.updateResourceInAppRoute(updatedResource);
-                var jsonObject = new JSONObject();
+                final var jsonObject = new JSONObject();
                 jsonObject.put("connectorResponse", response);
                 jsonObject.put("resourceID", resourceId.toString());
                 return ResponseEntity.ok(jsonObject.toJSONString());
-            }else{
-                return ResponseEntity.status(HttpStatus.NOT_FOUND).body(String.format("No resource with ID %s was found!", resourceId));
+            } else {
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body(String.format("No " +
+                        "resource with ID %s was found!", resourceId));
             }
         } catch (IOException e) {
+            log.error(e.getMessage(), e);
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(e.getMessage());
         }
     }
